requirements.txt
ansible>=4.0.0
dnspython>=2.0.0 